Rename `workspace` commands
