+++ cli/src/commands/workspace/list.test.ts
@@ -2,7 +2,7 @@ import { Config } from '@oclif/core';
 import fetch from 'node-fetch';
 import { afterEach, beforeEach, describe, expect, test, vi } from 'vitest';
 import { clearEnvVariables } from '../utils.test.js';
-import WorkspacesCreate from './create.js';
+import WorkspaceList from './list.js';
 
 vi.mock('node-fetch');
 
@@ -18,18 +18,7 @@ afterEach(() => {
 
 const fetchMock = fetch as unknown as ReturnType<typeof vi.fn>;
 
-describe('workspaces create', () => {
-  test('fails if the workspace name is not provided', async () => {
-    const config = await Config.load();
-    const list = new WorkspacesCreate([], config);
-
-    await expect(list.run()).rejects.toMatchInlineSnapshot(`
-      [Error: Missing 1 required arg:
-      workspace  The new workspace name
-      See more help with --help]
-    `);
-  });
-
+describe('workspaces list', () => {
   test('fails if the HTTP response is not ok', async () => {
     fetchMock.mockReturnValue({
       ok: false,
@@ -39,38 +28,48 @@ describe('workspaces create', () => {
     });
 
     const config = await Config.load();
-    const list = new WorkspacesCreate(['hello world'], config);
+    const list = new WorkspaceList([], config);
 
     await expect(list.run()).rejects.toThrow('Something went wrong');
 
     expect(fetchMock).toHaveBeenCalledOnce();
     expect(fetchMock.mock.calls[0][0]).toEqual('https://api.xata.io/workspaces');
-    expect(fetchMock.mock.calls[0][1].method).toEqual('POST');
+    expect(fetchMock.mock.calls[0][1].method).toEqual('GET');
   });
 
-  test.each([[false], [true]])('performs the creation with JSON enabled = %o', async (json) => {
+  test.each([[false], [true]])('returns the data with enabled = %o', async (json) => {
     fetchMock.mockReturnValue({
       ok: true,
       json: async () => ({
-        id: 'hello-world-1234'
+        workspaces: [
+          {
+            name: 'test',
+            id: 'test-1234',
+            role: 'Maintainer'
+          }
+        ]
       })
     });
 
     const config = await Config.load();
-    const list = new WorkspacesCreate(['hello world'], config);
+    const list = new WorkspaceList([], config);
 
-    expect(WorkspacesCreate.enableJsonFlag).toBe(true);
+    expect(WorkspaceList.enableJsonFlag).toBe(true);
     vi.spyOn(list, 'jsonEnabled').mockReturnValue(json);
 
-    const log = vi.spyOn(list, 'log');
+    const printTable = vi.spyOn(list, 'printTable');
 
     const result = await list.run();
 
     if (json) {
       expect(result).toMatchInlineSnapshot(`
-        {
-          "id": "hello-world-1234",
-        }
+        [
+          {
+            "id": "test-1234",
+            "name": "test",
+            "role": "Maintainer",
+          },
+        ]
       `);
     } else {
       expect(result).toBeUndefined();
@@ -78,12 +77,27 @@ describe('workspaces create', () => {
 
     expect(fetchMock).toHaveBeenCalledOnce();
     expect(fetchMock.mock.calls[0][0]).toEqual('https://api.xata.io/workspaces');
-    expect(fetchMock.mock.calls[0][1].method).toEqual('POST');
+    expect(fetchMock.mock.calls[0][1].method).toEqual('GET');
 
-    expect(log).toHaveBeenCalledTimes(json ? 0 : 1);
+    expect(printTable).toHaveBeenCalledTimes(json ? 0 : 1);
 
     if (!json) {
-      expect(log.mock.calls[0][0]).toEqual('✔ Workspace hello-world-1234 successfully created');
+      expect(printTable.mock.calls[0]).toMatchInlineSnapshot(`
+        [
+          [
+            "Name",
+            "Id",
+            "Role",
+          ],
+          [
+            [
+              "test",
+              "test-1234",
+              "Maintainer",
+            ],
+          ],
+        ]
+      `);
     }
   });
 });
