+++ cli/src/commands/workspaces/list.ts
@@ -1,4 +1,5 @@
 import { BaseCommand } from '../../base.js';
+import WorkspaceList from '../workspace/list.js';
 
 export default class WorkspacesList extends BaseCommand<typeof WorkspacesList> {
   static description = 'List your workspaces';
@@ -13,14 +14,12 @@ export default class WorkspacesList extends BaseCommand<typeof WorkspacesList> {
 
   static enableJsonFlag = true;
 
+  static hidden = true;
+
   async run(): Promise<any> {
-    const xata = await this.getXataClient();
-    const workspacesList = await xata.api.workspaces.getWorkspacesList();
+    this.warn('This command is deprecated. Please use `xata workspace list` instead.');
 
-    if (this.jsonEnabled()) return workspacesList.workspaces;
-
-    const headers = ['Name', 'Id', 'Role'];
-    const rows = workspacesList.workspaces.map((b) => [b.name, b.id, b.role]);
-    this.printTable(headers, rows);
+    const { argv } = await this.parseCommand();
+    return await WorkspaceList.run([...argv]);
   }
 }
